Simplify logging enabled by updates in pipe_base
The creation of a python logger to pass to piff with forwarding to lsst.log
is no longer necessary here thanks to recent updates in pipe_base (DM-30301).

The addition of _DefaultName ensures the log message will include the class
name (i.e. to be able to clearly identify where the message is coming from).