Add option for a per-band spatialOrder dict config
This is motivated by the fact that u band will almost always have
fewer PSF canditates than the other bands (throughput, stellar SED,
etc.), and that u-band coadds are highly unlikely to be used in
lensing/shear analyses. Setting a smaller spatial order for u
precludes an attempt at a higher order fit that inevitably won't
be well-constrained (and currently piff does not handle this internally
and returns a dodgy PSF model).  LSSTComCam overrides are being added
to obs_lsst, and other cameras should override in their obs_ package
as desired.  Wtih not override, the fallback is to the original
spatialOrder config.